% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval}
\alias{crossval}
\title{Uses cross-validation to find a good weighting of regularization}
\usage{
crossval(y, x, partitions = 5, degree = 1, lambda = c(0.1, 1, 10))
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Independent variable(s) in array form}

\item{partitions}{Number of partitions to divide the data into (n for n-fold validation)}

\item{degree}{Degree of polynomial to fit}

\item{lambda}{Vector of lambda weightings to try}
}
\value{
A list of the the best \code{lambda} value, and its fit (includes coefficients and error).
}
\description{
Uses cross-validation to find a good weighting of regularization
}
\details{
Splits the data into the number of partitions specified then attempts a fit on each partition, and calculates the error of the fit on the remaining data.
}
\examples{
crossval((1:10)^3, 1:10, partitions = 20, degree = 3, lambda = c(0.001, 0.01, 0.1, 1, 10))
}
