% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regfit.R
\name{reg.lm}
\alias{reg.lm}
\title{Fits a polynomial of specified degree to the data with a regularization term}
\usage{
reg.lm(y, x, degree = 1, lambda = 1)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Independent variable(s) in array form}

\item{degree}{Degree of polynomial to fit}

\item{lambda}{Weight of regularization term (may be selected using crossval)}
}
\value{
List with the coefficients for each term in the polynomial and the error.
}
\description{
Fits a polynomial of specified degree to the data with a regularization term
}
\details{
Finds the polynomial of specified degree that minimizes

\code{({y}-p({x}))^2+lambda*sum(coeff^2)}

Where \code{p} is the polynomial and \code{coeff} is its coefficients.

This function is useful when concerns about overfitting are present. \code{lambda} may be chosen by the \code{crossval} function.
}
\examples{
reg.lm((1:10)^3, 1:10, degree = 3, lambda = 1)
}
